Fix search results losing onclick. Set threshold for lazy loading
A threshold of 0px is now specified to the lazy loader, which is
better than the 500px default which would load more images than
is sometimes needed. This means the image definitely has to be
within the viewport before it is loaded.

Secondly, fixed the bug where some card results would sometimes lose
their ability to open the modal. The issue here is that when an
element is removed from the DOM, its onclick handler is also deleted.
There seems to be no way to re-bind this handler again even if the
element is re-appended to the DOM. This situation happens when the
resultsbox is emptied, and one of the new results is contained in
searchresults. Fortunately, the fix is easy - we can add an onclick
listener to parent elements that we know for certain will remain
on the page. In this case, we use the resultsbox div. Clicking on
a resultcard now propagates the click event upwards to the resultsbox.
The resultsbox itself is listening for clicks on article elements, and
when such an event is triggered, it will call selectedSong, passing
the element's ID. For this to work properly, we do now need to add the
song ID as an attribute to the article elements. This is also much
better now as we only need one onclick handler, and no need to
constantly append new ones.