Add another collection to retain upcoming queue
This is needed as if number of singers is modified during a session
and the browser is refreshed then the requests in the upcoming queue may
fragmented or pushed back to the awaiting queue. Having this second
collection to keep track of which items have fulfilled the singer
requirements before allows us to keep the upcoming queue intact.
This requires rewriting a key part of the getqueue logic to make
use of this new collection.

In addition, it's possible that when the number of singers requirement
is reduced, some of the awaiting songs can be added to the upcoming queue.
In order to make sure this is reflected, we need to update the
client.singers value and also tell the display to
rebuild the partialQueue divs. This is done by added a extra mode
parameter to the broadcast data.