Add duplicate songs setting to admin panel
A toggle switch has been added to the admin panel that allows controlling
whether duplicate songs should be allowed in the queue. This is accomplished by
new endpoint admin/allowdupes. This new control will allow the admin to
turn off duplicate songs during a live session. When the setting is toggled,
the awaiting queue will be refreshed accordingly. Note, however, this will not
affect the upcoming songs queue.