Update Heroku deploy script and Heroku dockerfile
Heroku no longer supports mLab which means we can't rely on Heroku to
automatically provision a mongoDB instance for us anymore. Instead, the
user will have to manually supply the mongoDB database address and
database name. The heroku deploy script has been updated to do this.

In this commit, the Dockerfile for Heroku deployment has also been updated so
that it no longer downloads from the upstream repo. This makes it easier
to deploy and test local changes. As part of the change, the dockerfile
has been moved to the root directory as currently only the directory the
Dockerfile is located is can be used as part of the build. In order to
avoid conflict with the normal Dockerfile, it has been renamed, and the
docker-compose.yml file also explicitly specifies the Dockerfile to use.