Add Cache-Control and validation to response headers
Add Cache-Control to HTTP response headers for search
and cover art requests. This will tell the client
browser to cache its responses for one day so that
subsequent requests can use the browser cache directly
and not need to make a server request. This will help
reduce the amount of data that needs to be sent from
between the server to the client and hopefully also reduce
the amount of searching on the database.

Last-Modified header is also added to the request endpoint
to validate whether a cached resource is still fresh,
and likewise an ETag header is attached the the cover
art response header. The Last-Modified header uses
a new field in the State collection "SongsLastUpdated"
to retrieve the last update of the Songs collection.
If this field does not exist then it will insert the
the server's initialisation time.
The ETag uses a sha256 hash of the image.

Also:
- Updated mongodb driver to v1.4.1