Fix lazy loading
Lazy loading currently has issues where images will fail to load
sometimes. Switch to native lazy loading using the loading="lazy"
attribute. For browsers that do not support this yet, use a lazy
loading polyfill. For browsers that do not support lazy loading, setting the
src attribute of img elements will cause it to fetch it straight away. In order
to lazy load, first check if lazy loading is supported and if not set the url
to data-lazy-src.  Afterwards, the polyfill script will move it to src when
necessary. Also tweak polyfill script to enable us to call it after adding the
new HTML elements. Also:
- Move package.json files inside static/frontend so that we don't need to keep
  copying it into the folder
- increase context deadline to account for long queries