Fix issue where requests do not go above 2 singers
In display.js we create div ids for the waiting items by
joining the queue item's request IDs via a '.' character. However
later we try to use jQuery to remove divs based on these IDs
which fail as it doesn't recognise the dots. This results in multiple
cards being shown for the same song. This is resolved not joining
the ids with a '.' as it's not needed anyway.